Rewrite SliderField
